# contracts/platform-v1/openapi.yaml
openapi: 3.1.0
info:
  title: Insight Platform API
  version: 1.0.0
  description: >-
    Current insight.platform/v1 public contract. Production L4-L6 qualification remains a
    separately reported deployment gate and does not change this repository contract status.
...
servers:
  - url: /v1
paths:
...
  /model-configuration:
    get:
      operationId: getModelConfiguration
      x-insight-operation-kind: query
      x-insight-command-receipt: false
      x-insight-permission: model.read
      x-insight-maximum-request-bytes: 16384
      responses:
        "200":
          description: Current authenticated model configuration result
          headers:
            Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
          content:
            application/json:
              schema: {$ref: "./schemas/model-configuration.schema.json#/$defs/ModelConfigurationCatalogViewV2"}
        "400": {$ref: "#/components/responses/PrivateApiProblem"}
        "401": {$ref: "#/components/responses/PrivateApiProblem"}
        "403": {$ref: "#/components/responses/PrivateApiProblem"}
        "404": {$ref: "#/components/responses/PrivateApiProblem"}
        "409": {$ref: "#/components/responses/PrivateApiProblem"}
        "503": {$ref: "#/components/responses/PrivateApiProblem"}
...
  /model-credentials/{secret_binding_id}:
    get:
      operationId: getModelCredential
      summary: Read safe current model credential metadata
      tags: [Resources]
      x-insight-permission: secret.bind OR secret.revoke
      parameters:
        - name: secret_binding_id
          in: path
          required: true
          schema: {$ref: "#/components/schemas/ModelCredentialId"}
      responses:
        "200":
          description: Current tenant-bound metadata without provider locators or credential material.
          headers:
            ETag: {schema: {type: string, maxLength: 128}}
            Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ModelCredentialMetadataViewV1"}
        "400": {$ref: "#/components/responses/PrivateApiProblem"}
        "401": {$ref: "#/components/responses/PrivateApiProblem"}
        "403": {$ref: "#/components/responses/PrivateApiProblem"}
        "404": {$ref: "#/components/responses/PrivateApiProblem"}
        "503": {$ref: "#/components/responses/PrivateApiProblem"}
  /model-credentials/{secret_binding_id}:revoke:
    post:
      operationId: revokeModelCredential
      summary: Revoke the current logical binding with generation and version CAS
      tags: [Resources]
      x-insight-permission: secret.revoke
      x-insight-idempotency: tenant_principal_scoped_receipt
      parameters:
        - name: secret_binding_id
          in: path
          required: true
          schema: {$ref: "#/components/schemas/ModelCredentialId"}
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RevokeModelCredentialRequestV1"}
      responses:
        "200":
          description: Current metadata after authorized CAS or Receipt replay; no physical provider deletion.
          headers:
            ETag: {schema: {type: string, maxLength: 128}}
            Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ModelCredentialMetadataViewV1"}
        "400": {$ref: "#/components/responses/PrivateApiProblem"}
        "401": {$ref: "#/components/responses/PrivateApiProblem"}
        "403": {$ref: "#/components/responses/PrivateApiProblem"}
        "404": {$ref: "#/components/responses/PrivateApiProblem"}
        "409": {$ref: "#/components/responses/PrivateApiProblem"}
        "412": {$ref: "#/components/responses/PrivateApiProblem"}
        "428": {$ref: "#/components/responses/PrivateApiProblem"}
        "503": {$ref: "#/components/responses/PrivateApiProblem"}
  /model-default:
    get:
      operationId: getModelDefault
      summary: Read the tenant-owned exact default Model Deployment pointer
      tags: [Resources]
      x-insight-permission: model.read
      responses:
        "200":
          description: Current Tenant version and configured default, including an explicit null when unset.
          headers:
            ETag: {schema: {type: string, maxLength: 128}}
            Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ModelDefaultViewV1"}
        "401": {$ref: "#/components/responses/PrivateApiProblem"}
        "403": {$ref: "#/components/responses/PrivateApiProblem"}
        "503": {$ref: "#/components/responses/PrivateApiProblem"}
    put:
      operationId: setModelDefault
      summary: Set or explicitly clear the tenant default under its current ETag
      tags: [Resources]
      x-insight-permission: model.write
      x-insight-idempotency: tenant_principal_scoped_receipt
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SetModelDefaultRequestV1"}
      responses:
        "200":
          description: The current Tenant projection after applying or replaying the authorized command.
          headers:
            ETag: {schema: {type: string, maxLength: 128}}
            Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ModelDefaultViewV1"}
        "400": {$ref: "#/components/responses/PrivateApiProblem"}
        "401": {$ref: "#/components/responses/PrivateApiProblem"}
        "403": {$ref: "#/components/responses/PrivateApiProblem"}
        "404": {$ref: "#/components/responses/PrivateApiProblem"}
        "409": {$ref: "#/components/responses/PrivateApiProblem"}
        "412": {$ref: "#/components/responses/PrivateApiProblem"}
        "428": {$ref: "#/components/responses/PrivateApiProblem"}
        "503": {$ref: "#/components/responses/PrivateApiProblem"}
...
components:
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 255, pattern: "^[ -~]+$"}
  responses:
    PrivateApiProblem:
      description: A bounded authenticated management problem without sensitive values.
      headers:
        trace-id: {$ref: "#/components/headers/TraceId"}
        Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiProblem"}
...
  headers:
    TraceId:
      required: true
      schema:
        $ref: "#/components/schemas/TraceId"
...
    PrivateNoStore:
      schema:
        type: string
        const: no-store, private, max-age=0
...
  schemas:
...
    ModelCredentialId:
      type: string
      pattern: "^sbd_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    ModelCredentialMetadataViewV1:
      type: object
      additionalProperties: false
      required: [schema_version, tenant_id, secret_binding_id, provider_id, purpose, state, generation, version, etag]
      properties:
        schema_version: {const: 1}
        tenant_id: {type: string, pattern: "^ten_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"}
        secret_binding_id: {$ref: "#/components/schemas/ModelCredentialId"}
        provider_id: {type: string, pattern: "^spr_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"}
        purpose: {const: model_api_key}
        state: {type: string, enum: [active, revoked]}
        generation: {type: integer, minimum: 1, maximum: 9223372036854775807}
        version: {type: integer, minimum: 1, maximum: 9223372036854775807}
        etag: {type: string, minLength: 1, maxLength: 128}
    RevokeModelCredentialRequestV1:
      type: object
      additionalProperties: false
      required: [schema_version, expected_generation]
      properties:
        schema_version: {const: 1}
        expected_generation: {type: integer, minimum: 1, maximum: 9223372036854775807}
...
    SetModelDefaultRequestV1:
      type: object
      additionalProperties: false
      required: [schema_version, default_model]
      properties:
        schema_version: {const: 1}
        default_model:
          oneOf:
            - {$ref: "#/components/schemas/ModelExactDeploymentRef"}
            - {type: "null"}
    ModelDefaultViewV1:
      type: object
      additionalProperties: false
      required: [schema_version, tenant_id, default_model, version, etag]
      properties:
        schema_version: {const: 1}
        tenant_id: {type: string, pattern: "^ten_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"}
        default_model:
          oneOf:
            - {$ref: "#/components/schemas/ModelExactDeploymentRef"}
            - {type: "null"}
        version: {type: integer, minimum: 1, maximum: 9223372036854775807}
        etag: {type: string, minLength: 1, maxLength: 128}
...
    ModelExactDeploymentRef:
      type: object
      additionalProperties: false
      required: [deployment_id, resource_kind, deployment_digest]
      properties:
        deployment_id:
          type: string
          pattern: "^mdep_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
        resource_kind: {const: model_deployment}
        deployment_digest: {$ref: "#/components/schemas/Digest"}
...
    Digest:
      $ref: ./schemas/nominal/digest.schema.json
...
    ApiProblem:
      $ref: ./schemas/nominal/api-problem.schema.json
    TraceId:
      $ref: ./schemas/nominal/trace-id.schema.json
...